+++ coreexample/src/androidTest/kotlin/me/proton/core/test/android/uitests/tests/medium/auth/signup/ExternalSetupTests.kt
@@ -53,7 +53,7 @@ class ExternalSetupTests : BaseTest() {
             .switchSignupType()
             .verify {
                 chooseUsernameElementsDisplayed()
-                suffixFilledButNotDisplayed(BuildConfig.HOST)
+                suffixNotDisplayed()
                 switchToExternalDisplayed()
             }