+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.1.9]
+
 ### Changes
 - MailSettingsRepository now supports offline changes.
 - Added ContactCard type 1 (Encrypted, but not signed).